Add translation to View. Move PlainTestMessage
