+++ src/main/java/client/interface_adapter/Main/MainController.java
@@ -4,17 +4,20 @@ import client.use_case.ReceiveMessage.ReceiveMessageInputBoundary;
 import client.use_case.ReceiveMessage.ReceiveMessageInteractor;
 import client.use_case.SendMessage.SendMessageInputBoundary;
 import client.use_case.SendMessage.SendMessageInputData;
+import client.use_case.Translate.TranslationInputBoundary;
+import client.use_case.Translate.TranslationInputData;
 
 public class MainController {
 
     private final SendMessageInputBoundary sendMessageInteractor;
-
     private final ReceiveMessageInputBoundary receiveMessageInteractor;
+    private final TranslationInputBoundary translationInteractor;
     private final String myUsername;
 
-    public MainController(String myUsername, SendMessageInputBoundary sendMessageInteractor, ReceiveMessageInputBoundary receiveMessageInteractor){
+    public MainController(String myUsername, SendMessageInputBoundary sendMessageInteractor, ReceiveMessageInputBoundary receiveMessageInteractor, TranslationInputBoundary translationInteractor){
         this.sendMessageInteractor = sendMessageInteractor;
         this.receiveMessageInteractor = receiveMessageInteractor;
+        this.translationInteractor = translationInteractor;
         this.myUsername = myUsername;
     }
 
@@ -25,6 +28,11 @@ public class MainController {
 
     public void getMessage(){
         receiveMessageInteractor.execute();
+    }
+
+    public String translateMessage(String message){
+        TranslationInputData in = new TranslationInputData(message);
+        return translationInteractor.execute(in);
     }
 
 }
