+++ src/main/java/client/use_case/Translate/TranslationInputBoundary.java
@@ -1,5 +1,5 @@
 package client.use_case.Translate;
 
 public interface TranslationInputBoundary {
-    void execute(TranslationInputData translationInputData);
+    String execute(TranslationInputData translationInputData);
 }
